+++ migrations/000001_create_subscriptions.sql
@@ -2,7 +2,7 @@
 CREATE TABLE subscriptions (
     id uuid PRIMARY KEY DEFAULT gen_random_uuid(),
     service_name TEXT NOT NULL,
-    price INTEGER NOT NULL CHECK (price > 0),
+    price BIGINT NOT NULL CHECK (price > 0),
     user_id uuid NOT NULL,
     started_at DATE NOT NULL,
     ended_at DATE,
